requirements.txt
aiohttp
aiosqlite
nextcord
akinator.py
mysql-connector